feat(cleanup,rbac): Ensure cleanup is setup before kube resources
This patch ensures that the cleanup is correctly built before applying
our Kubernetes resources. This will ensure that we can never be in a
state where the namespace exists, but our cleanup function has not run
yet.

This patch also ensures that the cleanup has time to run at the end with
a max timeout of 1 second on an interupt.